+++ stackoverflow-plugin-impl/src/main/java/com/victor/kochnev/plugin/stackoverflow/service/webclient/WebClientService.java
@@ -3,5 +3,5 @@ package com.victor.kochnev.plugin.stackoverflow.service.webclient;
 import com.victor.kochnev.plugin.stackoverflow.entity.StackOverflowQuestion;
 
 public interface WebClientService {
-    StackOverflowQuestion getStackOverflowInfo(Integer questionId);
+    StackOverflowQuestion getStackOverflowInfo(Long questionId);
 }
